Code changes to support the new Node.js support of websockets on OpenShift.
Really, this is a monkey patch that allows what I find to be the
easiest and most configurable approach to dealing with the need
for a potential port change in the near future. Plus, this code
is really not optimized, so I'm okay adding one additional HTTP
request via blocking script tag to the work flow.